refactor: move snapshot to util
