+++ util/snapshot/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "ckb-snapshot"
+version = "0.21.0-pre"
+license = "MIT"
+authors = ["Nervos Core Dev <dev@nervos.org>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ckb-types = { path = "../types" }
+ckb-chain-spec = { path = "../../spec" }
+ckb-store = { path = "../../store" }
+ckb-db = { path = "../../db" }
+ckb-traits = { path = "../../traits" }
+ckb-error = { path = "../../error" }
+ckb-proposal-table = { path = "../proposal-table" }
+# DO NOT UPDATE "im", until clippy#4121 be solved
+# https://github.com/rust-lang/rust-clippy/issues/4121
+im = "~12.3"
+arc-swap = "0.4"
